backend/requirements.txt
# Django Core
django
djangorestframework
python-dotenv
djangorestframework-simplejwt 
django-cors-headers

# PostgreSQL Connector
psycopg2-binary

# Celery for Asynchronous Tasks
celery
redis

# Document Parsing
PyMuPDF
python-docx
Markdown

# AI & Vector Database Clients
torch --index-url https://download.pytorch.org/whl/cpu
google-generativeai
sentence-transformers
pinecone

# Production Server
gunicorn
whitenoise

# Metrics Collection
django-prometheus